Fixed some memory leaks Fixed some cutout issues Fixed the status bar not getting destroyed on rom load Fixed a couple issues in the debugger code Configured Wabbitemu Release properly Fixed a bug that made save state loading impossible without virtualization on vista/7 Fixed the speed changing code, now doesn't use floats Fixed sending any non app Readded sending options on drag Added in flash writes for the 84p Fixed a bug in port 2 not identifying 84 series Added option to save state on exit

+++ debugger/dbreg.c
@@ -1105,7 +1105,7 @@ LRESULT CALLBACK RegProc(HWND hwnd, UINT Message, WPARAM wParam, LPARAM lParam) 
 			si.cbSize = sizeof(SCROLLINFO);
 			si.nMin = 0;
 			si.nMax = height;
-			si.nPos = percent * height;
+			si.nPos = (int) percent * height;
 			if (height != 0)
 				si.nPage = 40 - rc.bottom  / height;
 			SetScrollInfo(hwnd, SB_VERT, &si, TRUE);
